etl/requirements.txt
# Kafka
confluent-kafka[avro]==2.3.0
fastavro==1.9.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data processing
pandas==2.1.4

# Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilities
tqdm==4.66.1
faker==22.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
